[#523] Move system message filter to inline toggle in PanelHeader
Lift filter state from ChatPanel to ProjectDashboard so the toggle
renders in the same header row as the panel title. Removes the
separate filter row, saving vertical space. Toggle style matches
existing Auto/Sound patterns: "All ○" / "Agents ●".

Fixes #523
