+++ src/components/ChatPanel.tsx
@@ -126,28 +126,17 @@ function renderMessageWithMentions(text: string): React.ReactNode[] {
  */
 interface ChatPanelProps {
   projectId?: string;
+  /** #523: filter state lifted to parent so toggle renders in PanelHeader */
+  filterSystem?: boolean;
 }
 
-export default function ChatPanel({ projectId }: ChatPanelProps) {
-  return <ChatPanelAPI projectId={projectId} />;
+export default function ChatPanel({ projectId, filterSystem }: ChatPanelProps) {
+  return <ChatPanelAPI projectId={projectId} filterSystem={filterSystem} />;
 }
 
 /** API-driven fallback when iframe is blocked */
-function ChatPanelAPI({ projectId }: { projectId?: string }) {
+function ChatPanelAPI({ projectId, filterSystem = false }: { projectId?: string; filterSystem?: boolean }) {
   const channel = "general";
-  // #519: filter toggle — hide system/status messages. Persisted in
-  // localStorage (per-browser, not per-project). Default: OFF.
-  const [filterSystem, setFilterSystem] = useState(() => {
-    if (typeof window === "undefined") return false;
-    return localStorage.getItem("chatFilterSystem") === "1";
-  });
-  const toggleFilter = useCallback(() => {
-    setFilterSystem((prev) => {
-      const next = !prev;
-      localStorage.setItem("chatFilterSystem", next ? "1" : "0");
-      return next;
-    });
-  }, []);
   const [messages, setMessages] = useState<Message[]>([]);
   // #410: track whether the initial fetch has completed so we don't
   // flash the welcome/empty state while messages are still loading.
@@ -417,21 +406,6 @@ function ChatPanelAPI({ projectId }: { projectId?: string }) {
 
   return (
     <div className="flex flex-col h-full bg-bg">
-      {/* #519: filter toggle header */}
-      <div className="flex items-center justify-end h-7 px-3 shrink-0 border-b border-border">
-        <button
-          type="button"
-          onClick={toggleFilter}
-          title={filterSystem ? "Showing agent messages only — click to show all" : "Showing all messages — click to hide system/status noise"}
-          className={`px-1.5 py-0.5 text-[10px] border transition-colors ${
-            filterSystem
-              ? "border-accent/50 text-accent bg-accent/10 hover:bg-accent/20"
-              : "border-border text-text-muted hover:text-text hover:border-accent"
-          }`}
-        >
-          {filterSystem ? "Agents only" : "All messages"}
-        </button>
-      </div>
       {/* Auth error banner */}
       {authError && (
         <div className="px-3 py-2 bg-red-900/30 border-b border-red-700/50 text-[11px] text-red-400">
